+++ scala/sources/src/main/scala/io/cucumber/scala/Scenario.scala
@@ -15,32 +15,25 @@ import io.cucumber.core.backend.{Status, TestCaseState}
  */
 class Scenario(val delegate: TestCaseState) {
 
+  /**
+   * @return tags of this scenario.
+   */
   def getSourceTagNames: util.Collection[String] = delegate.getSourceTagNames
 
   /**
-   * Returns the current status of this scenario.
+   * Returns the current status of this test case.
    * <p>
-   * The scenario status is calculate as the most severe status of the
-   * executed steps in the scenario so far.
+   * The test case status is calculate as the most severe status of the
+   * executed steps in the testcase so far.
    *
-   * @return the current status of this scenario
+   * @return the current status of this test case
    */
   def getStatus: Status = Status.valueOf(delegate.getStatus.name)
 
   /**
-   * @return true if and only if { @link #getStatus()} returns "failed"
+   * @return true if and only if `getStatus` returns "failed"
    */
   def isFailed: Boolean = delegate.isFailed
-
-  @deprecated(message = "Use attach instead", since = "5.x")
-  def embed(data: Array[Byte], mediaType: String): Unit = {
-    delegate.embed(data, mediaType)
-  }
-
-  @deprecated(message = "Use attach instead", since = "5.7.0")
-  def embed(data: Array[Byte], mediaType: String, name: String): Unit = {
-    attach(data, mediaType, name)
-  }
 
   /**
    * Attach data to the report(s).
@@ -66,9 +59,16 @@ class Scenario(val delegate: TestCaseState) {
     delegate.attach(data, mediaType, name)
   }
 
-  @deprecated(message = "Use log instead", since = "5.7.0")
-  def write(text: String): Unit = {
-    log(text)
+  /**
+   * Attaches some text based data to the report.
+   *
+   * @param data      what to attach, for example html.
+   * @param mediaType what is the data?
+   * @param name      attachment name
+   * @see #attach(byte[], String, String)
+   */
+  def attach(data: String, mediaType: String, name: String): Unit = {
+    delegate.attach(data, mediaType, name)
   }
 
   /**